+++ api/src/communication/Communication.ts
@@ -12,7 +12,7 @@ class Communication {
             http,
             {
                 cors: {
-                    origin: 'http://localhost:3000',
+                    origin: 'http://127.0.0.1:3000',
                     methods: ['GET', 'POST']
                 }
             }