+++ web/src/@types/interfaces/Action.d.ts
@@ -2,5 +2,3 @@ interface Action<T> {
     type: string;
     payload?: T,
 }
-
-export default Action;