+++ web/src/@types/interfaces/Reducer.d.ts
@@ -1,0 +1,1 @@
+type Reducer<S> = (state: S, action: Action<any>) => S;