+++ web/src/@types/interfaces/client/IClientStates.d.ts
@@ -1,0 +1,3 @@
+interface IClientStates {
+    user: IUserState;
+}