+++ web/src/@types/interfaces/client/ObjectColor.d.ts
@@ -1,0 +1,1 @@
+type ObjectColor = { hex: number, alpha: number };