+++ web/src/@types/interfaces/client/index.d.ts
@@ -1,0 +1,5 @@
+import './states/IUserState';
+
+import './IClientStates';
+import './IUser';
+import './ObjectColor';