+++ web/src/@types/interfaces/client/states/IUserState.d.ts
@@ -1,8 +1,6 @@
-import IUser from "../IUser";
-
 interface IUserState {
     connected: boolean;
     user: IUser;
-}
 
-export default IUserState;
+    current_room: number;
+}