+++ web/src/@types/interfaces/index.d.ts
@@ -1,0 +1,4 @@
+import './client';
+
+import './Action';
+import './Reducer';