+++ web/src/data/reducers/Reducer.ts
@@ -1,5 +1,0 @@
-import Action from "../actions/Action";
-
-type Reducer<S> = (state: S, action: Action<any>) => S;
-
-export default Reducer;