+++ web/src/data/stores/clientStore.ts
@@ -1,9 +1,7 @@
-import mainReducer from '../reducers/client/mainReducer';
-import userReducer from '../reducers/client/userReducer';
-import { createStore, combineReducers } from 'redux';
+import userReducer from "../reducers/client/userReducer";
+import { createStore, combineReducers } from "redux";
 
 const clientStore = createStore(combineReducers({
-    main: mainReducer,
     user: userReducer,
 }));