+++ web/src/interfaces/client/states/IClientStates.ts
@@ -1,9 +1,0 @@
-import IMainState from "./IMainState";
-import IUserState from "./IUserState";
-
-interface IClientStates {
-    main: IMainState;
-    user: IUserState;
-}
-
-export default IClientStates;