+++ web/src/interfaces/client/states/IMainState.ts
@@ -1,9 +1,0 @@
-import { Shroom } from "@jankuss/shroom";
-import { Application } from "pixi.js";
-
-interface IMainState {
-    gameApplication: Application | null;
-    gameManager: Shroom | null;
-}
-
-export default IMainState;