+++ CHANGELOG.md
@@ -1,13 +1,17 @@
 # Unreleased
 
-## 17/11/2020
+## 23/12/2020
+- (@danielsolartech) (Web) Add hotel view, create room window and currency stats.
+- (@danielsolartech) (Web) Change to JS vanilla.
+
+## 17/12/2020
 - (@danielsolartech) Update shroom package version to 0.1.4.
 - (@danielsolartech) Add user info, currency and avatar packets.
 - (@danielsolartech) (Web) Add Redux to client states.
 - (@danielsolartech) (Web) Add websocket client.
 - (@danielsolartech) (API) Fix CORS with websocket connection.
 
-## 16/11/2020
+## 16/12/2020
 - (@danielsolartech) Add basic login.
 - (@danielsolartech) (API) Add HTTP and websocket server.
 - (@danielsolartech) (API) Add entities design.